requirements.txt
alembic==1.4.1
aniso8601==7.0.0
APScheduler==3.6.3
asgiref==3.2.8
astroid==2.4.1
Authlib==0.14.3
autopep8==1.5.2
branca==0.4.0
cachelib==0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
croniter==0.3.32
cryptography==2.9.2
Django==3.0.7
dominate==2.5.1
flake8==3.8.2
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-GraphQL==2.0.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
folium==0.10.1
idna==2.9
image==1.5.32
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
natsort==7.0.1
numpy==1.18.1
Pillow==7.1.2
pkg-resources==0.0.0
promise==2.3
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.2
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
zipp==3.1.0
